+++ src/parser/block/extension.test.ts
@@ -11,14 +11,15 @@ describe('Unit: parser/block/extension', () => {
       assert(!parser('~~~'));
       assert(!parser('~~~\n'));
       assert(!parser('~~~\na~~~'));
-      assert(!parser('~~~ a\n~~~'));
-      assert(!parser('~~~a b\n~~~'));
       assert(!parser(' ~~~\n~~~'));
     });
 
     it('valid', () => {
       assert(parser('~~~\n~~~'));
-      assert(parser('~~~a-b:c \n~~~'));
+      assert(parser('~~~a\n~~~'));
+      assert(parser('~~~a \n~~~'));
+      assert(parser('~~~a b \n~~~'));
+      assert(parser('~~~ a\n~~~'));
     });
 
   });